+++ lunes_cms/cms/models/discipline.py
@@ -1,11 +1,13 @@
 from django.db import models
 from django.db.models.deletion import CASCADE
 from django.contrib.auth.models import Group
+from django.contrib.contenttypes.fields import GenericRelation
 from django.utils.translation import ugettext_lazy as _
 
 from mptt.models import MPTTModel, TreeForeignKey
 
-from ..utils import get_child_count
+from ..utils import get_child_count, get_image_tag
+from .feedback import Feedback
 from .static import convert_umlaute_images
 
 
@@ -42,6 +44,7 @@ class Discipline(MPTTModel):
         related_name="children",
         verbose_name=_("parent"),
     )
+    feedback = GenericRelation(Feedback)
 
     def is_valid(self):
         """Checks if a discipline itself or one of its children has at least one training set.
@@ -68,6 +71,17 @@ class Discipline(MPTTModel):
                 "id", flat=True
             )
         return set(training_sets)
+
+    def image_tag(self):
+        """
+        Image thumbnail to display a preview of the icon
+
+        :return: img HTML tag to display an image thumbnail
+        :rtype: str
+        """
+        return get_image_tag(self.icon)
+
+    image_tag.short_description = ""
 
     def __str__(self):
         """String representation of Discipline instance