+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 UNRELEASED
 ----------
 
+* [ [#337](https://github.com/digitalfabrik/lunes-cms/issues/337) ] Remove image names
+* [ [#211](https://github.com/digitalfabrik/lunes-cms/issues/211) ] Add feedback endpoint to API
+* [ [#346](https://github.com/digitalfabrik/lunes-cms/issues/346) ] Add bulk action to mark feedback as read
+* [ [#317](https://github.com/digitalfabrik/lunes-cms/issues/317) ] Add icon preview for disciplines and training sets
+* [ [#350](https://github.com/digitalfabrik/lunes-cms/issues/350) ] Add counter for unread feedback entries
+
 
 2022.6.0
 --------